#14 move stuff to options.rs
