+++ src/lib.rs
@@ -4,14 +4,8 @@ mod config;
 mod transformer;
 mod printable_error;
 
-use serde_json::json;
-use std::env;
-
 use crate::worker::process;
-
-use crate::printable_error::PrintableError;
-use crate::config::parse_config_from_file;
-use crate::options::{Options, ErrorFormat, Variables};
+use crate::options::get_options;
 
 
 const HELP: &str = r#"
@@ -27,53 +21,6 @@ To get errors in JSON format for integration with other tools:
 
   cat original.csv | ysv configuration.yaml > result.csv
 "#;
-
-
-fn format_error(error: &PrintableError, format: &ErrorFormat) -> String {
-    match format {
-        ErrorFormat::HumanReadable => format!(
-            "ysv : Error {} | {}",
-            error.error_type,
-            error.error_description,
-        ),
-        ErrorFormat::JSON => json!(error).to_string()
-    }
-}
-
-// Fetch environment variables
-fn determine_variables() -> Variables {
-    let prefix = "YSV_VAR_";
-
-    env::vars().filter(
-        |(variable, _)| variable.starts_with(prefix)
-    ).map(
-        |(variable, value)| (
-            variable.replace(prefix, ""),
-            value,
-        )
-    ).collect()
-}
-
-
-fn get_options(args: Vec<String>) -> Result<Options, String> {
-    let error_format = ErrorFormat::JSON;
-    let first_argument = args.get(1);
-
-    if let None = first_argument {
-        return Err(String::from("Please provide configuration file path."))
-    }
-
-    let variables = determine_variables();
-
-    let config_result = parse_config_from_file(
-        first_argument.unwrap(),
-    );
-
-    match config_result {
-        Ok(config) => Ok(Options { error_format, config, variables }),
-        Err(err) => Err(format_error(&err, &error_format))
-    }
-}
 
 
 // Did the user request the built-in help message?
