+++ src/main.rs
@@ -1,8 +1,6 @@
-use std::env;
-// use crate::lib::run;
-
-
 extern crate ysv;
+
+use std::env;
 use ysv::run;
 
 fn main() -> Result<(), String> {
